Use 32-bit values for signal fences
The HSA model uses 64-bit on all shipped platforms, but the motivation
here is that the `wait_mem` only accepts a 32-bit pointer for PM4. Since
I made the questionalbe choice to go all-in on PM4, I shall die on this
hill. The memory is still 64-bit backed, so it's actually valid to use
it with HSA interop if such a thing every occurred, (lol)

+++ include/libkfd/context.h
@@ -95,7 +95,7 @@ public:
 
   // Register a handler with the context's watcher threads.
   std::expected<void, Error> register_handler(Signal &sig, Condition cond,
-                                              uint64_t value, SignalHandler cb,
+                                              uint32_t value, SignalHandler cb,
                                               void *user_data);
   std::expected<void, Error> register_handler(FaultHandler cb, void *user_data);
 
@@ -105,7 +105,7 @@ private:
   friend class Signal;
 
   std::expected<uint64_t *, Error> event_slot(uint32_t id);
-  std::expected<uint64_t *, Error> fence_slot(uint32_t id);
+  std::expected<uint32_t *, Error> fence_slot(uint32_t id);
 
   std::expected<void, Error> register_handler(Event &event, SignalHandler cb,
                                               void *user_data = nullptr);
